refactor: enhance packaging
