+++ src/main/kotlin/org/shangahi/sellio_backend/api/controller/DiscountController.kt
@@ -1,7 +1,7 @@
 package org.shangahi.sellio_backend.api.controller
 
-import org.shangahi.sellio_backend.api.dto.DiscountResponse
-import org.shangahi.sellio_backend.api.dto.PageResponse
+import org.shangahi.sellio_backend.api.dto.response.DiscountResponse
+import org.shangahi.sellio_backend.api.dto.response.PageResponse
 import org.shangahi.sellio_backend.service.DiscountService
 import org.springframework.data.domain.Pageable
 import org.springframework.data.web.PageableDefault
